Tests: Extract table-post-fixer tests.
